add Routing via state management
the various views of the game are related to state variables, ie. 'gameOver' being true renders the Score component. CurrentImage being changed renders that image as the current game.
these state variables are changed to force the desired component to render. For example: from the level selection component (the landing page), clicking the highscore link will mark gameOver as true, and set currentImage to 1 (as a default). Clicking Levels from the Scores component removes the currentImage & imageID states, and sets gameOver to false, thus forcing the LevelSelection component to render.